Update to v3.1.2:
 - Fixed: Within Campaign Cartographer, I override the Dispose() method for my ModHost, so I can dispose of my custom chat commands. However, I didn't then add in the base.Dispose(); to bubble down to the host. This, in turn, meant that none of the Gantry stuff was disposed of; including everything within the IOC container; all the paths for that specific world; settings; everything.

+++ src/ApacheTech.VintageMods.CampaignCartographer/Features/PlayerPins/Dialogue/PlayerPinsDialogue.cs
@@ -174,7 +174,7 @@ namespace ApacheTech.VintageMods.CampaignCartographer.Features.PlayerPins.Dialog
 
         private bool OnColourChanged(ColourChannel channel, int value)
         {
-            PlayerPinHelper.Colour = PlayerPinHelper.Colour.UpdateColourChannel(channel, value);
+            PlayerPinHelper.Colour = PlayerPinHelper.Colour.UpdateColourChannel(channel, (byte)value);
             RefreshValues();
             return true;
         }
